% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hccPIRS.R
\name{hccPIRS}
\alias{hccPIRS}
\title{Replication stress-related prognostic index in HCC}
\usage{
hccPIRS(
  expr = NULL,
  scaleFlag = FALSE,
  centerFlag = FALSE,
  doplot = TRUE,
  fig.path = getwd(),
  fig.name = "heatmap of replication stress",
  enrich = "gsva",
  width = 1,
  height = 4
)
}
\arguments{
\item{expr}{A numeric expression matrix with row features and sample columns.}

\item{scaleFlag}{A logic value to indicate if the expression data should be further scaled. FALSE by default.}

\item{centerFlag}{A logic value to indicate if the expression data should be further centered. TRUE by default.}

\item{doplot}{A logic value to indicate whether to generate heatmap of replication stress signatures and PIRS score; FALSE by default.}

\item{fig.path}{A string value to indicate the output path for storing the heatmap.}

\item{fig.name}{A string value to indicate the name of the heatmap.}

\item{enrich}{A string value to indicate the method for single-sample enrichment analysis. Allowed values contain c('gsva', 'ssgsea', 'zscore', 'plage'); 'gsva' by default.}

\item{width}{A numeric value to indicate the relative width for each cell in the heatmap; 1 by default.}

\item{height}{A numeric value to indicate the relative height for each cell in the heatmap; 4 by default.}
}
\value{

}
\description{
This function calculates replication stress-related prognostic index (PIRS) for HBV-associated HCC patients, and estimates the enrichment of 21 replication stress signatures.
}
\examples{
library(hccPIRS)
load(system.file("extdata", "tpm.demo.RData", package = "hccPIRS", mustWork = TRUE)) # load example data
res <- hccPIRS(expr = tpm.demo,
               scaleFlag  = FALSE,
               centerFlag = FALSE,
               doplot = TRUE,
               fig.path = getwd(),
               fig.name   = "heatmap of replication stress",
               enrich = "gsva",
               width = 6,
               height = 4)
pirs <- res$pirs # extract normalized PIRS score for each sample
print(pirs)
rsMat <- res$RS.score # extract enrichment score for replication stress signatures
rsMat[1:21, 1:3]
res$hm # show the heatmap
}
\references{
Dreyer, SB, Upstill-Goddard, R, Paulus-Hock, V, Paris, C, Lampraki, E-M, Dray, E, et al. (2021). Targeting DNA Damage Response and Replication Stress in Pancreatic Cancer. Gastroenterology 160: 362-377.e313.
}
\author{
Xiaofan Lu
}
