% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{pirs.coeff}
\alias{pirs.coeff}
\title{Cox proportional hazards regression results of replication stress-related genes}
\format{
A data frame with 9 rows (replication stress related genes) and 8 statistics
\describe{contains coefficient to calculate replication stress-related prognostic index.}
}
\usage{
pirs.coeff
}
\description{
A data frame storing feature coefficient for replication stress-related genes.
}
\keyword{datasets}
