% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{RS.signature}
\alias{RS.signature}
\title{Replication stress signatures}
\format{
A list for 21 signatures
\describe{contains replication stress related genes in a total of 21 signatures for enrichment analysis.}
}
\usage{
RS.signature
}
\description{
A list storing 21 replication stress signatures.
}
\keyword{datasets}
